1. matrix_example03, matrix_example04 프로젝트 추가 2. 프로젝트 타겟 설정을 x86에서 x64로 변경 3. 행,열 타입을 size_t에서 col_t, row_t로 세분화 및 통일 4. 각 example 프로젝트의 타입 통일

+++ matrix_vc/include/matrix_error.h
@@ -66,7 +66,7 @@ bool		ErrMsg::setErrString		(	const char		string[]	///< 에러 메시지
 
 	do
 	{
-		int		length		=	strnlen(string, MAX_ERR_STRING_LEN);
+		size_t	length		=	strnlen(string, MAX_ERR_STRING_LEN);
 		if( length == 0 )
 		{
 			break;
